+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_dynamic/avoid_dynamic_rule.dart
@@ -2,28 +2,23 @@
 
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/ast/visitor.dart';
-import 'package:analyzer/dart/element/element.dart';
-import 'package:analyzer/dart/element/type.dart';
-import 'package:collection/collection.dart';
 
-import '../../../../../utils/flutter_types_utils.dart';
 import '../../../../../utils/node_utils.dart';
 import '../../../lint_utils.dart';
 import '../../../models/internal_resolved_unit_result.dart';
 import '../../../models/issue.dart';
 import '../../../models/severity.dart';
-import '../../models/flutter_rule.dart';
+import '../../models/common_rule.dart';
 import '../../rule_utils.dart';
 
 part 'visitor.dart';
 
-class AlwaysRemoveListenerRule extends FlutterRule {
-  static const String ruleId = 'always-remove-listener';
+class AvoidDynamicRule extends CommonRule {
+  static const String ruleId = 'avoid-dynamic';
 
-  static const _warningMessage =
-      'Listener is not removed. This might lead to a memory leak.';
+  static const _warning = 'Avoid using dynamic type.';
 
-  AlwaysRemoveListenerRule([Map<String, Object> config = const {}])
+  AvoidDynamicRule([Map<String, Object> config = const {}])
       : super(
           id: ruleId,
           severity: readSeverity(config, Severity.warning),
@@ -32,18 +27,18 @@ class AlwaysRemoveListenerRule extends FlutterRule {
 
   @override
   Iterable<Issue> check(InternalResolvedUnitResult source) {
-    final _visitor = _Visitor();
+    final visitor = _Visitor();
 
-    source.unit.visitChildren(_visitor);
+    source.unit.visitChildren(visitor);
 
-    return _visitor.missingInvocations
-        .map((invocation) => createIssue(
+    return visitor.nodes
+        .map((node) => createIssue(
               rule: this,
               location: nodeLocation(
-                node: invocation,
+                node: node,
                 source: source,
               ),
-              message: _warningMessage,
+              message: _warning,
             ))
         .toList(growable: false);
   }