+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_async_await/visitor.dart
@@ -1,0 +1,19 @@
+part of 'prefer_async_await_rule.dart';
+
+class _Visitor extends RecursiveAstVisitor<void> {
+  final _invocations = <MethodInvocation>[];
+
+  Iterable<MethodInvocation> get invocations => _invocations;
+
+  @override
+  void visitMethodInvocation(MethodInvocation node) {
+    super.visitMethodInvocation(node);
+
+    final target = node.realTarget;
+
+    if ((target?.staticType?.isDartAsyncFuture ?? false) &&
+        node.methodName.name == 'then') {
+      _invocations.add(node);
+    }
+  }
+}