+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * fix: add `monorepo` flag for `check-unused-code` command.
+* fix: ignore a class usage inside `State<T>` for `check-unused-code` command.
+* fix: correctly handle variables declaration for `check-unused-code` command.
+* feat: add static code diagnostics `avoid-dynamic`, `prefer-async-await`.
 
 ## 4.10.1